Use Kafka messages to update missions tab instead of polling
Currently we use polling in the missions tab. This makes the server side of the console listen for Kafka messages and send messages from topics (specified through an env var) through a socket.io connection.

This means that when websockets are available, there'll be a real-time stream of information about current missions for the responder and location updates instead of using polling or a separate simulation.

This does not touch the dashboard. Only the mission tab is changed.